+++ us_somo/develop/src/us_hydrodyn_core.cpp
@@ -167,8 +167,8 @@ static void outward_translate_2_spheres(float *r1, // radius of sphere 1
 
    float sfact =
       sqrt(
-           (*r1 + *r2) * (*r1 + *r2) *
-           ((*r1 * *r1 * (-1 + v1[0] * v1[0] + v1[1] * v1[1] + v1[2] * v1[2]) -
+           (double)(*r1 + *r2) * (*r1 + *r2) *
+           ((double)(*r1 * *r1 * (-1 + v1[0] * v1[0] + v1[1] * v1[1] + v1[2] * v1[2]) -
 
              2 * *r1 * *r2 * (1 + v1[0] * v2[0] + v1[1] * v2[1] + v1[2] * v2[2]) +
 
@@ -7466,7 +7466,7 @@ double US_Hydrodyn::total_volume_of_bead_model( vector < PDB_atom > &bead_model 
    {
       if ( bead_model[ i ].active ) 
       {
-         tot_vol += bead_model[ i ].bead_computed_radius * bead_model[ i ].bead_computed_radius * bead_model[ i ].bead_computed_radius;
+         tot_vol += (double) bead_model[ i ].bead_computed_radius * bead_model[ i ].bead_computed_radius * bead_model[ i ].bead_computed_radius;
       }
    }
    return tot_vol * pi43;
